added readme; removed unaltered demo file from root and put it in test dir instead to avoid any confusion
