+++ test/index.html
@@ -4,9 +4,9 @@
     <meta charset="utf-8">
     <title>Templatizer</title>
     <link rel="stylesheet" href="qunit.css">
-    <script src="../demo_output.js"></script>
+    <script src="demo_output_no_mixins.js"></script>
     <script>window.templatizer_unaltered = templatizer;</script>
-    <script src="../demo_output_mixins.js"></script>
+    <script src="../demo_output.js"></script>
     <script src="qunit.js"></script>
     <script src="tests.js"></script>
 </head>
